Added levelup & jackpot bonus Added selfpromo require level

+++ cogs/selfpromo.py
@@ -4,6 +4,7 @@ import re
 from discord.ext import commands
 
 from config import Channel
+from models import MemberModel
 
 class SelfPromo(commands.Cog):
     def __init__(self, client: commands.Bot):
@@ -13,15 +14,19 @@ class SelfPromo(commands.Cog):
     @commands.Cog.listener('on_message')
     async def promote_message_handler(self, message: discord.Message):
         if message.channel.id == Channel.SELFPROMO and message.author.bot == False:
-            urls = re.findall("(?P<url>https?://[^\s]+)", message.content)
-            if len(urls) > 0:
-                await message.delete()
-                promo_message = await message.channel.send(urls[0])
-                if self.reply_message:
-                    await self.reply_message.delete()
-                em = discord.Embed(description=f'**Can\'t send any message?** [Click here for more information](https://discord.com)', color=discord.Colour.purple())
-                self.reply_message = await promo_message.reply(embed=em)
-                await promo_message.delete(delay=43200)
+            member = await MemberModel.find_one(MemberModel.member_id == message.author.id)
+            if member.level >= 5:
+                urls = re.findall("(?P<url>https?://[^\s]+)", message.content)
+                if len(urls) > 0:
+                    await message.delete()
+                    promo_message = await message.channel.send(urls[0])
+                    if self.reply_message:
+                        await self.reply_message.delete()
+                    em = discord.Embed(description=f'**Can\'t send any message?** [Click here for more information](https://discord.com)', color=discord.Colour.purple())
+                    self.reply_message = await promo_message.reply(embed=em)
+                    await promo_message.delete(delay=43200)
+                else:
+                    await message.delete()
             else:
                 await message.delete()
